Move NotNull annotation to annotation package
