--- chunky/src/java/se/llbit/util/NotNull.java
@@ -1,5 +1,0 @@
-package se.llbit.util;
-
-public @interface NotNull {
-
-}
