+++ chunky/src/java/se/llbit/util/annotation/NotNull.java
@@ -1,0 +1,5 @@
+package se.llbit.util.annotation;
+
+public @interface NotNull {
+
+}
